requirements.txt
fastapi==0.115.5
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
passlib==1.7.4
PyJWT==2.10.0
pymongo==4.10.1
uvicorn==0.32.1


